pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bitinformation"
version = "0.1.0"
authors = [
    {name = "Eugen Betke", email = "eugen.betke@ecmwf.int"}, 
]
description = "Bitwise information content computation"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
]
dependencies = [
    "scipy>=1.9.3",
    "numpy>=1.23.1",
    "pandas>=1.5.1",
    "pyyaml>=6.0",
]
license = "GPL-3.0-or-later"

[project.urls]
"Homepage" = "https://git.ecmwf.int/users/maee/repos/bitinformation/browse"

